updated omni types and converting between colors
