+++ src/bebone/assets/assets.h
@@ -1,6 +1,6 @@
 #ifndef _BEBONE_ASSETS_ASSETS_H_
 #define _BEBONE_ASSETS_ASSETS_H_
 
-#include "image/image.h"
+#include "image/image.tpp"
 
 #endif
