+++ src/bebone/gfx/opengl/opengl_texture.cpp
@@ -24,31 +24,6 @@ namespace bebone::gfx::opengl {
         glBindTexture(textureType, 0);
     }
 
-    GLTexture::GLTexture(const std::shared_ptr<Image>& image, const GLenum& textureType, const GLenum& format, const GLenum& pixelType)
-        : textureType(textureType), width(image->get_width()), height(image->get_height()) {
-
-        // int colorChannelNumber;
-
-        // stbi_set_flip_vertically_on_load(true);
-
-        // unsigned char* bytes = stbi_load(image, &width, &height, &colorChannelNumber, 0);
-
-        glGenTextures(1, &id);
-        glBindTexture(textureType, id);
-
-        glTexParameteri(textureType, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-        glTexParameteri(textureType, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-
-        glTexParameteri(textureType, GL_TEXTURE_WRAP_S, GL_NEAREST);
-        glTexParameteri(textureType, GL_TEXTURE_WRAP_T, GL_NEAREST);
-
-        glTexImage2D(textureType, 0, GL_RGBA, width, height, 0, format, pixelType, image->data());
-        glGenerateMipmap(textureType);
-
-        // stbi_image_free(bytes);
-        glBindTexture(textureType, 0);
-    }
-
     GLTexture::~GLTexture() {
         destroy();
     }
